Added an implementation for Wiring random() function
This function uses the True Random Number generator from inside the STM32F746
For normal pseudo-random numbers as in Arduino the chora built-in functions can be used.

Also added a demo project with a signal generator showing a random noise generated using
the random() function both in time and frequency.